// src/server/api/models/banking/models/AccountHold.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { AccountHoldStatus } from './AccountHoldStatus';
import type { AccountHoldType } from './AccountHoldType';
import type { LocalDateTime } from './LocalDateTime';
import type { SubAccountType } from './SubAccountType';
import type { UUID } from './UUID';

export type AccountHold = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The client ID from the crm service
     */
    clientId: UUID;
    /**
     * The client description
     */
    clientLabel: string;
    /**
     * The client code from CRM
     */
    clientCode: string;
    /**
     * The account ID that the hold applies to
     */
    accountId: UUID;
    /**
     * The account label
     */
    accountLabel: string;
    /**
     * The account number
     */
    accountNo: string;
    /**
     * The sub account type
     */
    subAccountType: SubAccountType;
    /**
     * A unique code generated by the system
     */
    code: string;
    /**
     * The hold amount
     */
    amount: number;
    /**
     * The transaction response code
     */
    responseCode?: string | null;
    /**
     * The transaction response message
     */
    responseMessage?: string | null;
    /**
     * The account currency
     */
    currency: string;
    /**
     * The hold description
     */
    label: string;
    /**
     * The status of the hold
     */
    status: AccountHoldStatus;
    /**
     * The type of the hold
     */
    holdType: AccountHoldType;
    /**
     * The transaction notes
     */
    notes?: string | null;
    /**
     * The expiration time date
     */
    expirationTime: LocalDateTime;
    /**
     * The idempotency key
     */
    idemPotencyId?: UUID | null;
};